Added buildoff logic
AI is kind of stupid and is more likely to build itself off (and lose) than win a game. Hence, I added some simple recognition if there is no available moves for the current player. This also ends the game. For the long run, this might be too expensive, but right now, it's fine.